repeater is working now

+++ src/Controller/RepeaterController.php
@@ -12,6 +12,7 @@ use App\Service\RoomAddService;
 use App\Service\ServerUserManagment;
 use App\Service\UserService;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\DependencyInjection\ParameterBag\ParameterBagInterface;
 use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -23,43 +24,46 @@ class RepeaterController extends AbstractController
     /**
      * @Route("/room/repeater/new", name="repeater_new")
      */
-    public function index(Request $request, TranslatorInterface $translator, RepeaterService $repeaterService): Response
+    public function index(ParameterBagInterface $parameterBag, RoomAddService $roomAddService, Request $request, TranslatorInterface $translator, RepeaterService $repeaterService): Response
     {
         //todo check if allowed
-        $room = $this->getDoctrine()->getRepository(Rooms::class)->find($request->get('room'));
 
+        $room = $this->getDoctrine()->getRepository(Rooms::class)->find($request->get('room'));
         $repeater = new Repeat();
         $form = $this->createForm(RepeaterType::class, $repeater, ['action' => $this->generateUrl('repeater_new', ['room' => $room->getId()])]);
 
-        try {
+//        try {
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $repeater = $form->getData();
+            if($repeater->getRepetation() > $parameterBag->get('laf_max_repeat') ){
+                $snack = $translator->trans('Sie dürfen nur maximal {amount} Wiederholungen angeben',array('{amount}'=>$parameterBag->get('laf_max_repeat') ));
+                return $this->redirectToRoute('dashboard', array('snack' => $snack, 'color' => 'danger'));
+            }
             $em = $this->getDoctrine()->getManager();
             foreach ($room->getUser() as $data) {
                 $room->addPrototypeUser($data);
-            }
-            $em->persist($room);
-            $em->flush();
-            $repeater->setPrototyp($room);
-            $repeater->setStartDate($room->getStart());
-            $em->persist($repeater);
-            $em->flush();
-            $repeaterService->createNewRepeater($repeater);
-            foreach ($room->getUser() as $data) {
                 $room->removeUser($data);
             }
             $em->persist($room);
             $em->flush();
+
+            $repeater->setPrototyp($room);
+            $repeater->setStartDate($room->getStart());
+            $em->persist($repeater);
+            $em->flush();
+            $userAttributes = $repeater->getPrototyp()->getUserAttributes()->toArray();
+            $repeater = $repeaterService->createNewRepeater($repeater);
+            $roomAddService->addUserRepeat($repeater);
             $snack = $translator->trans('Sie haben Erfolgreich einen Serientermin erstellt');
             return $this->redirectToRoute('dashboard', array('snack' => $snack, 'color' => 'success'));
         }
 
-        } catch (\Exception $exception) {
-            $snack = $translator->trans('Fehler, Bitte kontrollieren Sie ihre Daten.');
-            return $this->redirectToRoute('dashboard', array('snack' => $snack, 'color' => 'danger'));
-        }
+//        } catch (\Exception $exception) {
+//            $snack = $translator->trans('Fehler, Bitte kontrollieren Sie ihre Daten.');
+//            return $this->redirectToRoute('dashboard', array('snack' => $snack, 'color' => 'danger'));
+//        }
         return $this->render('repeater/index.html.twig', [
             'form' => $form->createView(),
         ]);
@@ -68,40 +72,47 @@ class RepeaterController extends AbstractController
     /**
      * @Route("/room/repeater/edit/repeat", name="repeater_edit_repeater")
      */
-    public function editRepeater(Request $request, TranslatorInterface $translator, RepeaterService $repeaterService): Response
+    public function editRepeater(ParameterBagInterface  $parameterBag, Request $request, TranslatorInterface $translator, RepeaterService $repeaterService, RoomAddService $roomAddService): Response
     {
         //todo check if allowed
         $repeater = $this->getDoctrine()->getRepository(Repeat::class)->find($request->get('repeat'));
         $form = $this->createForm(RepeaterType::class, $repeater, ['action' => $this->generateUrl('repeater_edit_repeater', ['repeat' => $repeater->getId()])]);
 
-        try {
+//        try {
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $repeater = $form->getData();
+            if($repeater->getRepetation() > $parameterBag->get('laf_max_repeat') ){
+                $snack = $translator->trans('Sie dürfen nur maximal {amount} Wiederholungen angeben',array('{amount}'=>$parameterBag->get('laf_max_repeat') ));
+                return $this->redirectToRoute('dashboard', array('snack' => $snack, 'color' => 'danger'));
+            }
             $em = $this->getDoctrine()->getManager();
-
             foreach ($repeater->getRooms() as $data) {
+                foreach ($data->getUser() as $data2) {
+                    $data2->removeRoom($data);
+                    $em->persist($data2);
+                }
+                foreach ($data->getUserAttributes() as $data2) {
+                    $em->remove($data2);
+                    $data->removeUserAttribute($data2);
+                }
                 $em->remove($data);
+                $repeater->removeRoom($data);
             }
             $em->persist($repeater);
             $em->flush();
-
-            $repeaterService->createNewRepeater($repeater);
-            $room = $repeater->getPrototyp();
-            foreach ($room->getUser() as $data) {
-                $room->removeUser($data);
-            }
-            $em->persist($room);
-            $em->flush();
+            $userAttributes = $repeater->getPrototyp()->getUserAttributes()->toArray();
+            $repeater = $repeaterService->createNewRepeater($repeater);
+            $roomAddService->addUserRepeat($repeater);
             $snack = $translator->trans('Sie haben Erfolgreich einen Serientermin bearbeitet');
             return $this->redirectToRoute('dashboard', array('snack' => $snack, 'color' => 'success'));
         }
 
-        } catch (\Exception $exception) {
-            $snack = $translator->trans('Fehler, Bitte kontrollieren Sie ihre Daten.');
-            return $this->redirectToRoute('dashboard', array('snack' => $snack, 'color' => 'danger'));
-        }
+//        } catch (\Exception $exception) {
+//            $snack = $translator->trans('Fehler, Bitte kontrollieren Sie ihre Daten.');
+//            return $this->redirectToRoute('dashboard', array('snack' => $snack, 'color' => 'danger'));
+//        }
         return $this->render('repeater/index.html.twig', [
             'form' => $form->createView(),
         ]);
@@ -115,17 +126,15 @@ class RepeaterController extends AbstractController
         //todo check if allowed
         $repeater = $this->getDoctrine()->getRepository(Repeat::class)->find($request->get('repeat'));
         $em = $this->getDoctrine()->getManager();
-
         foreach ($repeater->getRooms() as $data) {
-            $em->remove($data);
+            foreach ($data->getUser() as $data2) {
+                $data->removeUser($data2);
+                $em->persist($data);
+            }
         }
-        $prototype = $repeater->getPrototyp();
+
         $repeater->setPrototyp(null);
         $em->persist($repeater);
-        $em->flush();
-        $em->remove($repeater);
-        $em->flush();
-        $em->remove($prototype);
         $em->flush();
         $snack = $translator->trans('Sie haben Erfolgreich einen Serientermin gelöscht');
         return $this->redirectToRoute('dashboard', array('snack' => $snack, 'color' => 'success'));
@@ -134,26 +143,25 @@ class RepeaterController extends AbstractController
     /**
      * @Route("/room/repeater/edit/room", name="repeater_edit_room")
      */
-    public function editPrototype(Request $request, UserService $userService,TranslatorInterface $translator, RepeaterService $repeaterService, ServerUserManagment $serverUserManagment): Response
+    public function editPrototype(RoomAddService $roomAddService, Request $request, UserService $userService, TranslatorInterface $translator, RepeaterService $repeaterService, ServerUserManagment $serverUserManagment): Response
     {
         //todo check if allowed
         $servers = $serverUserManagment->getServersFromUser($this->getUser());
         $room = $this->getDoctrine()->getRepository(Rooms::class)->find($request->get('id'));
-        $repeater = $room->getRepeater();
         $form = $this->createForm(RoomType::class, $room, ['server' => $servers, 'action' => $this->generateUrl('repeater_edit_room', ['id' => $room->getId()])]);
         $form->add('repeaterRemoved', CheckboxType::class, array(
             'label' => 'label.repeaterRemoved',
             'required' => false,
             'translation_domain' => 'form'));
-        try {
+        //    try {
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $em = $this->getDoctrine()->getManager();
             $room = $form->getData();
-            if($room->getRepeaterRemoved()){
-                foreach ($room->getUser() as $data){
-                    $userService->editRoom($data,$room);
+            if ($room->getRepeaterRemoved()) {
+                foreach ($room->getUser() as $data) {
+                    $userService->editRoom($data, $room);
                 }
                 $em->persist($room);
                 $em->flush();
@@ -161,37 +169,16 @@ class RepeaterController extends AbstractController
                 return $this->redirectToRoute('dashboard', array('snack' => $snack, 'color' => 'success'));
             }
 
-            foreach ($room->getUser() as $data) {
-                $room->addPrototypeUser($data);
+            $repeaterService->replaceRooms($room);
 
-            }
-            $em->persist($room);
-            $em->flush();
-
-            $proOld = $repeater->getPrototyp();
-            $proOld->setRepeater(null);
-            $em->persist($proOld);
-            $proDub = clone $room;
-            $em->persist($proDub);
-            $room->setRepeater(null);
-            $repeater->setPrototyp($room);
-            $em->persist($repeater);
-            $em->flush();
-
-            $repeaterService->replaceRooms($repeater, $room);
-            foreach ($room->getUser() as $data) {
-                $room->removeUser($data);
-            }
-            $em->persist($room);
-            $em->flush();
             $snack = $translator->trans('Sie haben Erfolgreich einen Serientermin bearbeitet');
             return $this->redirectToRoute('dashboard', array('snack' => $snack, 'color' => 'success'));
         }
 
-        } catch (\Exception $exception) {
-            $snack = $translator->trans('Fehler, Bitte kontrollieren Sie ihre Daten.');
-            return $this->redirectToRoute('dashboard', array('snack' => $snack, 'color' => 'danger'));
-        }
+//        } catch (\Exception $exception) {
+//            $snack = $translator->trans('Fehler, Bitte kontrollieren Sie ihre Daten.');
+//            return $this->redirectToRoute('dashboard', array('snack' => $snack, 'color' => 'danger'));
+//        }
         return $this->render('base/__newRoomModal.html.twig', [
             'form' => $form->createView(),
             'title' => $translator->trans('Serienelement bearbeiten')
